feat(file-system): 数据侧栏可收藏，展开折叠不再拖着表格重绘
侧栏拆成独立组件，展开/收藏只更新这一栏。收藏区和聊天一样用星星，点过的表会出现在上方收藏里。
